fixed routes & add element switcher
