+++ client/public/locales/en/translation.json
@@ -8,6 +8,6 @@
     "head" : "Sing In",
     "email": "email",
     "password": "password",
-    "button": "SING IN"
+    "button": "Send"
   }
 }
