+++ client/src/components/custom-routes/custom-routes.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Layout from "../../pages/Layout";
-import SingIn from "../../pages/SingIn/SingIn";
-import NoPage from "../../pages/NoPage/NoPage";
+import Layout from "pages/Layout";
+import SingIn from "pages/SingIn/SingIn";
+import NoPage from "pages/NoPage/NoPage";
 
 const CustomRoutes = () => {
   return (
