+++ client/src/components/header/header.jsx
@@ -1,8 +1,6 @@
 import './header.scss'
-import Cookies from "js-cookie";
-import i18n from "i18next";
-import { useState } from "react"
 import { useTranslation } from "react-i18next";
+import Switcher from 'components/switch-language/switch-language'
 
 const Header = () => {
   const { t } = useTranslation();
@@ -16,20 +14,6 @@ const Header = () => {
       path: "#",
     },
   ];
-
-  const languages = [
-    { name: "English", code: "en" },
-    { name: "Українська", code: "uk" },
-  ];
-
-  const currentLocale = Cookies.get("i18next") || "en";
-  const [language, setLanguage] = useState(currentLocale);
-
-  const handleChangeLocale = (e) => {
-    const lang = e.target.value;
-    setLanguage(lang);
-    i18n.changeLanguage(lang);
-  };
 
   return (
     <header>
@@ -45,16 +29,7 @@ const Header = () => {
           </ul>
         </nav>
         <div className="spacer"></div>
-        <div className="switcher">
-          <span>{t("header.language")}</span>{" "}
-          <select onChange={handleChangeLocale} value={language}>
-            {languages.map(({ name, code }) => (
-              <option key={code} value={code}>
-                {name}
-              </option>
-            ))}
-          </select>
-        </div>
+        <Switcher/>
       </div>
     </header>
   );
