+++ client/src/i18n.js
@@ -3,10 +3,6 @@ import { initReactI18next } from "react-i18next";
 import Backend from "i18next-http-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
 
-i18n.on('languageChanged', function (lng) {
-  localStorage.setItem("lng", lng);
-})
-
 i18n
   .use(Backend)
   .use(LanguageDetector)
@@ -14,8 +10,8 @@ i18n
   .init({
     fallbackLng: "en",
     detection: {
-      order: ["path", "localStorage", "htmlTag", "cookie"],
-      caches: ["localStorage", "cookie"],
+      order: ["cookie"],
+      caches: ["cookie"],
     },
   });
