+++ client/src/pages/Layout.js
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from '../components/header/header'
+import Header from 'components/header/header'
 
 const Layout = () => {
   return (
